Finish converting Map<Target, count> state to separate arrays
