+++ src/app/Archives/Archives.tsx
@@ -40,7 +40,6 @@ import { ServiceContext } from '@app/Shared/Services/Services';
 import { Card, CardBody, EmptyState, EmptyStateIcon, Tab, Tabs, Title } from '@patternfly/react-core';
 import { SearchIcon } from '@patternfly/react-icons';
 import { AllTargetsArchivedRecordingsTable } from './AllTargetsArchivedRecordingsTable';
-import { AllTargetsArchivedRecordingsTableV2 } from './AllTargetsArchivedRecordingsTableV2';
 import { BreadcrumbPage } from '@app/BreadcrumbPage/BreadcrumbPage';
 import { ArchivedRecordingsTable } from '@app/Recordings/ArchivedRecordingsTable';
 import { Target } from '@app/Shared/Services/Target.service';
@@ -68,8 +67,7 @@ export const Archives = () => {
       </>);
     }
     return (<>
-      {/* <AllTargetsArchivedRecordingsTable /> */}
-      <AllTargetsArchivedRecordingsTableV2 />
+      <AllTargetsArchivedRecordingsTable />
     </>);
   }, [archiveEnabled]);
